requirements.txt
qbittorrent-api
transmission-rpc
python-dotenv
